+++ tlv/include/tlv/tlv.h
@@ -19,6 +19,7 @@
 #include "tlv/reader/reader.h"
 #include "tlv/reader/walker.h"
 #include "tlv/reader/scanner.h"
+#include "tlv/query/query.h"
 #include "tlv/writer/writer.h"
 #include "tlv/schemas/schema.h"
 #include "tlv/config.h"